requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
et_xmlfile==2.0.0
mysqlclient==2.2.7
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
